requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
bleach==5.0.1
Brotli==1.0.9
crispy-bulma==0.8.0
dj-database-url==1.0.0
Django==4.1.1
django-crispy-forms==1.14.0
django-markdownify==0.9.2
gunicorn==20.1.0
importlib-metadata==5.0.0
Markdown==3.4.1
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
webencodings==0.5.1
whitenoise==6.2.0
zipp==3.10.0
